+++ src/.vscode/tasks.json
@@ -8,13 +8,13 @@
             "args": [
                 //"-g",
                 "-std=c++17",
-                "./Rectangle_Detection/main.cpp",
-                "./Detect_Rectangles/ContourRectangles.hpp",
-                "./Detect_Rectangles/ContourRectangles.cpp",
-                "./Detect_Rectangles/GaussianBlurTrackbar.hpp",
-                "./Detect_Rectangles/CannyThresholdTrackbar.hpp",
-                "./Detect_Rectangles/GaussianBlurTrackbar.cpp",
-                "./Detect_Rectangles/CannyThresholdTrackbar.cpp",
+                "main.cpp",
+                "./ModularClasses/ContourRectangles.hpp",
+                "./ModularClasses/ContourRectangles.cpp",
+                "./ModularClasses/GaussianBlurTrackbar.hpp",
+                "./ModularClasses/CannyThresholdTrackbar.hpp",
+                "./ModularClasses/GaussianBlurTrackbar.cpp",
+                "./ModularClasses/CannyThresholdTrackbar.cpp",
                 //"./Rectangle_Detection/RectDetect1.hpp",
                 //"./Rectangle_Detection/RectDetect1.cpp",
                 //"find_paper.hpp",
