--- src/apply_contour.hpp
@@ -1,11 +1,0 @@
-#ifndef __APPLY_CONTOUR__
-#define __APPLY_CONTOUR__
-
-#include "find_paper.hpp"
-
-/*----Canny trackbar functions----*/
-extern void adjust_contour_trackbar_pos(int, void* );
-extern void apply_Contours();
-extern void init_contour_trackbar();
-
-#endif
